// src/models/member.model.ts
/*TODO: add missed properties*/
export interface Member {
  id: string;
  cin: string;
  nom: string;
  prenom: string;
  dateNaissance: Date;
  email: string ;
  cv: string;
  photo: string;
  dateInscription: string;
  sujet: string;
  diplome: string;
  encadrant: any;
  etablissement: string;
  grade: string;
  encadrant_id: any;
}